# re-use the [4]byte for getting blocksize.