this gives about a 2.5% speed bump.

```
benchmark           old ns/op     new ns/op     delta
BenchmarkRead-4     6454267       6272807       -2.81%

benchmark           old allocs     new allocs     delta
BenchmarkRead-4     34020          30684          -9.81%
```